backend/requirements.txt
# Voice Sentiment Analysis Backend
# Python dependencies for Flask + Whisper + Transformers

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# API & Validation
pydantic>=2.5.0
python-multipart>=0.0.6

# Machine Learning - Whisper (Speech-to-Text)
openai-whisper>=20231106

# Machine Learning - Transformers (Sentiment Analysis)
torch>=2.0.0
transformers>=4.30.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.11.0

# Configuration & Environment
python-dotenv>=1.0.0

# Async Task Queue (for Phase 2)
celery>=5.3.0
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Logging & Monitoring
structlog>=23.2.0

# Gunicorn for production
gunicorn>=21.2.0

